requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.4.6
billiard==3.3.0.23
celery==3.1.23
colorama==0.3.7
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
drfdocs==0.0.11
kombu==3.0.35
lazy-object-proxy==1.2.2
Markdown==2.6.6
MySQL-python==1.2.5
pkg-resources==0.0.0
pylint==1.5.6
pytz==2016.4
redis==2.10.5
six==1.10.0
wrapt==1.10.8
